Move HighlightsCalculator to common SDK
