+++ MapboxSearch/sdk-common/api/api-metalava.txt
@@ -16,6 +16,32 @@ package com.mapbox.search.common {
     property public final com.mapbox.search.common.AsyncOperationTask COMPLETED;
   }
 
+  public interface DistanceCalculator {
+    method public double distance(com.mapbox.geojson.Point from, com.mapbox.geojson.Point to);
+    method public default static double distanceOnSphere(com.mapbox.geojson.Point from, com.mapbox.geojson.Point to);
+    method public default static com.mapbox.search.common.DistanceCalculator instance(double latitude);
+    field public static final com.mapbox.search.common.DistanceCalculator.Companion Companion;
+  }
+
+  public static final class DistanceCalculator.Companion {
+    method public double distanceOnSphere(com.mapbox.geojson.Point from, com.mapbox.geojson.Point to);
+    method public com.mapbox.search.common.DistanceCalculator instance(double latitude);
+  }
+
+  public final class DistanceCalculatorKt {
+  }
+
+  public interface HighlightsCalculator {
+    method public default static com.mapbox.search.common.HighlightsCalculator getInstance();
+    method public java.util.List<kotlin.Pair<java.lang.Integer,java.lang.Integer>> highlights(String name, String query);
+    field public static final com.mapbox.search.common.HighlightsCalculator.Companion Companion;
+  }
+
+  public static final class HighlightsCalculator.Companion {
+    method public com.mapbox.search.common.HighlightsCalculator getInstance();
+    property public final com.mapbox.search.common.HighlightsCalculator INSTANCE;
+  }
+
   @kotlinx.parcelize.Parcelize public final class IsoCountryCode implements android.os.Parcelable {
     ctor public IsoCountryCode(String code);
     method public String getCode();
