+++ MapboxSearch/sdk-common/build.gradle
@@ -104,5 +104,6 @@ dependencies {
 
     testImplementation "org.junit.jupiter:junit-jupiter-api:$junit_5_version"
     testImplementation "org.junit.jupiter:junit-jupiter-engine:$junit_5_version"
+    testImplementation "org.junit.jupiter:junit-jupiter-params:$junit_5_version"
     testImplementation project(":common-tests")
 }
